refactor: break AspectsSidebar into components, setup proper linting
