+++ .eslintignore
@@ -2,3 +2,4 @@ coverage/*
 dist/
 node_modules/
 jest.config.js
+babel.config.js
