+++ package.json
@@ -14,8 +14,8 @@
     "preinstall": "npm run build",
     "build": "fedx-scripts babel -x .js,.jsx,.ts,.tsx src --out-dir dist --ignore **/*.test.jsx,**/*.test.js,**/setupTest.js && cp src/*.css dist/",
     "i18n_extract": "fedx-scripts formatjs extract",
-    "lint": "fedx-scripts eslint --ext .js --ext .jsx .",
-    "lint:fix": "fedx-scripts eslint --fix --ext .js --ext .jsx .",
+    "lint": "fedx-scripts eslint --ext .js --ext .jsx --ext .ts --ext .tsx .",
+    "lint:fix": "fedx-scripts eslint --fix --ext .js --ext .jsx --ext .ts --ext .tsx .",
     "test": "fedx-scripts jest --coverage --passWithNoTests"
   },
   "husky": {
@@ -58,9 +58,21 @@
   },
   "devDependencies": {
     "@edx/browserslist-config": "^1.1.1",
+    "@edx/eslint-config": "^4.3.0",
+    "@edx/typescript-config": "^1.1.0",
     "@openedx/frontend-build": "14.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.57.1",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "glob": "11.0.1",
     "husky": "9.1.7",
-    "jest": "29.7.0"
+    "jest": "29.7.0",
+    "typescript": "^4.9.5"
   }
 }
