+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@edx/typescript-config",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "dist"
+  },
+  "include": ["src/**/*", ".eslintrc.js"],
+  "exclude": ["dist", "node_modules"]
+}
